Split fuse code into rshim_fuse.c
This commit splits the fuse/cuse code into rshim_fuse.c. So fuse
is just one way to implement the boot/console/networking services.
Other user interfaces, like networking or CLI, could be added
when needed.

This commit also unifies the build process for Linux and FreeBSD,
and adds '--enable-fuse' in the configure options.
