--- Makefile.freebsd
@@ -1,52 +1,0 @@
-# SPDX-License-Identifier: GPL-2.0-only
-#
-# Copyright 2019 Mellanox Technologies. All Rights Reserved.
-#
-
-TARGET:= bfrshim
-
-PREFIX:=/usr/local
-LOCALBASE:=/usr/local
-BINDIR:=${PREFIX}/sbin
-LIBDIR:=${PREFIX}/lib
-INCDIR:=${PREFIX}/include
-MANDIR:=${PREFIX}/share/man
-CC:= cc
-INSTALL:= install
-
-CFLAGS = \
-       -O2 \
-       -I$(LOCALBASE)/include \
-       -I$(LOCALBASE)/include/libepoll-shim \
-       -D_FILE_OFFSET_BITS=64 \
-       -DHAVE_RSHIM_NET \
-       -DHAVE_RSHIM_CUSE \
-       -DHAVE_RSHIM_USB \
-       -DHAVE_RSHIM_PCIE \
-       -g
-
-LIBS = -L$(LOCALBASE)/lib \
-       -lpci \
-       -lepoll-shim \
-       -lusb \
-       -lcuse \
-       -lpthread
-
-.PHONY: $(TARGET) clean
-
-all: $(TARGET)
-
-OBJS := $(patsubst %.c,%.o,$(wildcard src/*.c))
-
-$(TARGET): $(OBJS)
-	$(CC) $^ -o $@ $(LIBS)
-
-%.o: %.c
-	$(CC) $(CFLAGS) -c -o $@ $<
-
-clean:
-	$(RM) $(TARGET) *.o
-
-install:
-	$(INSTALL) -d $(DESTDIR)$(BINDIR)
-	$(INSTALL) -m 755 $(TARGET) $(DESTDIR)$(BINDIR)/
