+++ README.md
@@ -19,8 +19,7 @@ internal rshim registers.
   Require FreeBSD 12.0+ with packages autoconf, automake, gmake, libepoll-shim,
   libpciaccess, libpci, pkgconf.
 
-  Follow the same steps as build in Linux, or use 'gmake -f Makefile.freebsd'
-  to build it. Use 'gmake install' to install it.
+  Follow the same steps above to build it. Use 'gmake install' to install it.
 
 *) Usage
